requirements.txt

    matplotlibpyplot==3.1.2
    numpy